+++ src/lib/forms-schema/elements/FileElement.ts
@@ -9,6 +9,8 @@ import {
   conditionallyShowSchemas,
 } from '../property-schemas'
 
+export const type = 'file'
+
 export default Joi.object({
   id,
   name,