+++ src/lib/forms-schema/elements/HTMLElement.ts
@@ -1,6 +1,9 @@
 import Joi from 'joi'
 import { id, name, label, conditionallyShowSchemas } from '../property-schemas'
 import { base64DataRegex } from '../common'
+
+export const type = 'html'
+
 export default Joi.object({
   id,
   name,